Move JAVA_HOME logic from swift-java's Package.swift to swift-jni
